Add custom icons and improve resource dict

+++ CurrentWeatherInfo.cs
@@ -38,7 +38,7 @@ namespace WeatherCustomControl
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
+        public Double pressure { get; set; }
         public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
